[Linux] Move compositor shader into its own GObject
Extracts the OpenGL shader program used by FlCompositorOpenGL into a new
FlCompositorOpenGLShader GObject. This encapsulates the shader sources,
program, uniform locations and vertex buffer, along with their
compilation and cleanup, behind a small API.
