+++ engine/src/flutter/shell/platform/linux/BUILD.gn
@@ -118,6 +118,7 @@ source_set("flutter_linux_sources") {
     "fl_binary_messenger.cc",
     "fl_compositor.cc",
     "fl_compositor_opengl.cc",
+    "fl_compositor_opengl_shader.cc",
     "fl_compositor_software.cc",
     "fl_dart_project.cc",
     "fl_display_monitor.cc",
